+++ CHANGELOG.md
@@ -7,7 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 # Unreleased
 
-- **added:** Implement `Hash` to `Value`
+- None.
+
+# 0.1.4 (08. February, 2023)
+
+- **added:** Implement `Hash` to `Value` ([#93])
+
+[#93]: https://github.com/EmbarkStudios/mirror-mirror/pull/93
 
 # 0.1.3 (08. February, 2023)